Added code to split out argv lists of command line
This function iterates over the string splitting it by newlines, and then re-iterates over each line, splitting it by spaces. It then populates an argv list for the bhyve command and for the loader, respectively.

Since this file reuses part of the code of qemu_parse_command.c (in bhyveCommandLine2argv), add the appropriate copyright notices.