Support custom bootloaders
Also treat grub-bhyve as a custom bootloader, since we don't get the device map
in the native format anyways, we can't reconstruct the complete boot order
anyways. While it is possible to check what type the grub boot disk is
by checking if the --root argument is "cd" or "hd0,msdos1", and then just use
the first disk found, implementing the grub-bhyve argument parser as-is in the
grub-bhyve source would mean adding a dependency to argp or duplicating lots
of the code of argp. Therefore it's not really worth implementing that now.